tests/wpt/tests/html/semantics/embedded-content/the-canvas-element/toDataURL.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: toDataURL.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>toDataURL.default</h1>
<p class="desc">toDataURL with no arguments returns a PNG</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL with no arguments returns a PNG");
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL();
assert_regexp_match(data, /^data:image\/png[;,]/);


});
</script>
